requirements.txt
flask
python-dotenv
requests
scikit-learn
pandas
gunicorn